Move shared items to utils
